Step 1: Background service worker runs the content script, and receives its scraped data as a return value. Step 2: We save this value as a global inside the Background service worker (for now. But maybe we can save it in chrome.storage) Step 3: Since this value is now saved, it can be obtained from the frontend through Plasmo's messaging API.
 (I didded it)
