+++ package.json
@@ -11,6 +11,7 @@
     "package": "plasmo package"
   },
   "dependencies": {
+    "@plasmohq/messaging": "^0.1.6",
     "apexcharts": "^3.37.1",
     "next": "^13.1.5",
     "plasmo": "^0.67.3",
