+++ src/background/messages/getScrapeData.ts
@@ -1,0 +1,13 @@
+import type { PlasmoMessaging } from "@plasmohq/messaging"
+import type { CourseHeader, ShowCourseTabPayload } from ".../backgroundInterfaces";
+import {getScrapedCourseData} from "../index";
+ 
+const handler: PlasmoMessaging.MessageHandler = async (req, res) => {
+  const data = getScrapedCourseData();
+ 
+  res.send({
+    data
+  })
+}
+ 
+export default handler
